uint32_t instead of unsigned long
 * lots of fixes for 64bit <--> 32bit communication

+++ include/host.h
@@ -20,7 +20,7 @@
 typedef struct
 {
     char *name;
-    unsigned long address;
+    uint32_t address;
     int failed;
     double failuretime;
     int port;
